docker-images/application/requirements.txt
kafka-python
pyyaml
faker